requirements.txt
alabaster==0.7.12
AST==0.0.2
Babel==2.6.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
docopt==0.6.2
docutils==0.14
enum34==1.1.6
humanfriendly==4.17
idna==2.8
imagesize==1.1.0
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.10
MarkupSafe==1.1.0
monotonic==1.5
packaging==18.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
property-manager==2.3.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
traitlets==4.3.2
typing==3.6.6
urllib3==1.24.1
verboselogs==1.7
wcwidth==0.1.7
